Just a place to start
Simply makes the 2d board array (this can be accomplished in our Game class if we want to but I did it here to keep it simple for now. If you compile this with Critter.cpp on flip you get a print out of our board with 0s and 1s wherever **.placeCritter(x,y) has been called.